[스프링 - 협업 미션] 와이비(허영범) 미션 제출합니다. (#12)
* feat : 1차 배포 완료

* docs(ToDo) : 3단계 요금 조회 기능 문서

* test(FareTest) : 요금 생성 관련 테스트 작성

* feat(Fare): 거리에 따른 요금 계산 기능 구현

* feat(PathResponse) : 경로 응답에 요금 정보 추가

* docs(ToDo) : 4단계 요금 정책 추가 문서

* feat(schema.sql) : sql 변경

* feat(Fare) : 연령 및 노선을 고려한 추가 요금 구현

* feat: 가장 높은 금액의 추가 요금을 적용하는 기능 추가

* refactor: 기존 Line 관련 코드에 추가요금 필드 추가

* fix : 자바 의존성 문제 해결

* feat(AuthenticationMemberArgumentResolver) : 요금 관련 로그인 처리 추가

* refactor(LoginMember) : 전달 인자 타입 수정

* fix(fare): 요금 정책 오류 수정

* test(PathAcceptanceTest): 요금 정책 통합 테스트 추가

* feat(MapController) : Section 정보 포함한 전체 노선 목록 조회 기능 추가

* feat: 중복 email을 검사하는 기능 추가

* feat: 예외 처리 결과 프론트로 전달

* feat: swagger 추가

* fix(DataLoader) : 초기 데이터 설정 시에 DB검사

* fix(MemberService) : 이메일 중복 검사 오류 수정

* fix(SubwayControllerAdvice) : 예외 처리 규칙에 맞게 변경

* fix(DataLoader) : 초기 자료 설정 오류 수정

* fix(LineRequest) : 노선 추가 요금 입력 안할 시 오류 처리

* fix(StationDao) : 검증 비즈니스 로직을 Dao에서 분리

* fix : 로컬 환경 수정

* fix(ExistedStationOnPathException) : 노선 등록된 역 삭제 예외 메세지 변경

* fix(StationService) : 로그 제거

* refactor : 기본 도메인 메서드 위치 변경

* fix(LineService) : 역 삭제시에 구간 삭제 안되던 오류 수정

* polish : 컨벤션 및 불필요한 패키지 제거

* test(StationAcceptanceTest) : 구간 위 역 제거 오류 테스트

* feat(Station) : 이름 유효성 검사

* feat(Line) : 이름 유효성 검사

* fix(StationService) : @Transactional Annotation 추가

* fix(LineService) : @Transactional Annotationn 추가

* refactor(StationController) : Void 반환 수정

* refactor(PathRequest) : Command 객체 도입하여 값 검증

* refactor(PathService) : PathService 불필요한 예외 처리 제거

* test(StationService) : Station 통합 테스트 추가

* test(LineService) : Line 통합 테스트 추가

* refactor(AgeAppliedRule) : 패키지 이동

* refactor(DistanceAppliedRule) : Distance에 따른 요금 변경 적용 방법 변경

* feat(Guest) : 인증 안된 사용자를 위한 Guest 도입

* feat(LoginInterceptor) : 로그인 관련 Interceptor 도입

* refactor(AuthenticationMemberArgumentResolver) : Resolver 통합

+++ src/main/java/wooteco/subway/station/application/StationService.java
@@ -1,22 +1,25 @@
 package wooteco.subway.station.application;
 
+import java.util.List;
+import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import wooteco.subway.station.dao.StationDao;
 import wooteco.subway.station.domain.Station;
 import wooteco.subway.station.dto.StationRequest;
 import wooteco.subway.station.dto.StationResponse;
-
-import java.util.List;
-import java.util.stream.Collectors;
+import wooteco.subway.station.exception.ExistedStationOnPathException;
 
 @Service
 public class StationService {
+
     private StationDao stationDao;
 
     public StationService(StationDao stationDao) {
         this.stationDao = stationDao;
     }
 
+    @Transactional
     public StationResponse saveStation(StationRequest stationRequest) {
         Station station = stationDao.insert(stationRequest.toStation());
         return StationResponse.of(station);
@@ -30,11 +33,15 @@ public class StationService {
         List<Station> stations = stationDao.findAll();
 
         return stations.stream()
-                .map(StationResponse::of)
-                .collect(Collectors.toList());
+            .map(StationResponse::of)
+            .collect(Collectors.toList());
     }
 
+    @Transactional
     public void deleteStationById(Long id) {
+        if (stationDao.countStationOccurrenceOnPath(id) != 0) {
+            throw new ExistedStationOnPathException();
+        }
         stationDao.deleteById(id);
     }
 }
